+++ src/mocks/MockUNO.sol
@@ -9,8 +9,10 @@ contract MockUNO is ERC20 {
         _mint(msg.sender, 1000000 ether);
     }
 
-    /**@notice mints `amount` of mUNO and transfers it to msg.sender */
-    function getMockDAI(uint amount) public {
+    /**
+     * @notice mints `amount` of mUNO and transfers it to msg.sender
+     */
+    function getMockDAI(uint256 amount) public {
         _mint(msg.sender, amount);
     }
 }
